rtw89: Add device 0cb8:c549 to the USB ID table
Normally, Realtek BT devices are recognized by their device class, device sub class,
and device protocol values. Those three values for the problem device has are
0xe0, 0x01, and 0x01, which is the first entry in the device table. Despite having
these values, the  device is not recognized as belonging to this driver. Once the
generic device is recognized, then the blacklist table is scanned to find the
specific device.

As a temporary workaround, I added the explicit USB device to the first table.
This action "should" cause the driver to load.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>